// sources-frontend/src/app/modules/addresses/models/Address.ts
export interface Address {
  data_crypt?: string;
  symmetric_key_crypt?: string;
  id?: string
  addressFullName?: string
  area?: string
  city?: string
  street?: string
  house?: string
  frame?: string
  letters?: string
  stock?: string
  apartment?: string
  fio?: string
  phone?: string
  is_default?: any
  index?: string
  region?: string
  uniq_id?: string


}